update teams synchronization endpoints and house them in separate file
