+++ github/teams_test.go
@@ -956,127 +956,3 @@ func TestTeamsService_RemoveTeamProjectByName(t *testing.T) {
 		t.Errorf("Teams.RemoveTeamProjectByName returned error: %v", err)
 	}
 }
-
-func TestTeamsService_ListIDPGroupsInOrganization(t *testing.T) {
-	client, mux, _, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc("/orgs/o/team-sync/groups", func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, "GET")
-		testFormValues(t, r, values{
-			"page": "url-encoded-next-page-token",
-		})
-		fmt.Fprint(w, `{"groups": [{"group_id": "1",  "group_name": "n", "group_description": "d"}]}`)
-	})
-
-	opt := &ListCursorOptions{Page: "url-encoded-next-page-token"}
-	groups, _, err := client.Teams.ListIDPGroupsInOrganization(context.Background(), "o", opt)
-	if err != nil {
-		t.Errorf("Teams.ListIDPGroupsInOrganization returned error: %v", err)
-	}
-
-	want := &IDPGroupList{
-		Groups: []*IDPGroup{
-			{
-				GroupID:          String("1"),
-				GroupName:        String("n"),
-				GroupDescription: String("d"),
-			},
-		},
-	}
-	if !reflect.DeepEqual(groups, want) {
-		t.Errorf("Teams.ListIDPGroupsInOrganization returned %+v. want %+v", groups, want)
-	}
-}
-
-func TestTeamsService_ListIDPGroupsForTeam(t *testing.T) {
-	client, mux, _, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc("/orgs/o/teams/s/team-sync/group-mappings", func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, "GET")
-		fmt.Fprint(w, `{"groups": [{"group_id": "1",  "group_name": "n", "group_description": "d"}]}`)
-	})
-
-	groups, _, err := client.Teams.ListIDPGroupsForTeam(context.Background(), "o", "s")
-	if err != nil {
-		t.Errorf("Teams.ListIDPGroupsForTeam returned error: %v", err)
-	}
-
-	want := &IDPGroupList{
-		Groups: []*IDPGroup{
-			{
-				GroupID:          String("1"),
-				GroupName:        String("n"),
-				GroupDescription: String("d"),
-			},
-		},
-	}
-	if !reflect.DeepEqual(groups, want) {
-		t.Errorf("Teams.ListIDPGroupsForTeam returned %+v. want %+v", groups, want)
-	}
-}
-
-func TestTeamsService_CreateOrUpdateIDPGroupConnections(t *testing.T) {
-	client, mux, _, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc("/orgs/o/teams/s/team-sync/group-mappings", func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, "PATCH")
-		fmt.Fprint(w, `{"groups": [{"group_id": "1",  "group_name": "n", "group_description": "d"}]}`)
-	})
-
-	input := IDPGroupList{
-		Groups: []*IDPGroup{
-			{
-				GroupID:          String("1"),
-				GroupName:        String("n"),
-				GroupDescription: String("d"),
-			},
-		},
-	}
-
-	groups, _, err := client.Teams.CreateOrUpdateIDPGroupConnections(context.Background(), "o", "s", input)
-	if err != nil {
-		t.Errorf("Teams.CreateOrUpdateIDPGroupConnections returned error: %v", err)
-	}
-
-	want := &IDPGroupList{
-		Groups: []*IDPGroup{
-			{
-				GroupID:          String("1"),
-				GroupName:        String("n"),
-				GroupDescription: String("d"),
-			},
-		},
-	}
-	if !reflect.DeepEqual(groups, want) {
-		t.Errorf("Teams.CreateOrUpdateIDPGroupConnections returned %+v. want %+v", groups, want)
-	}
-}
-
-func TestTeamsService_CreateOrUpdateIDPGroupConnections_empty(t *testing.T) {
-	client, mux, _, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc("/orgs/o/teams/s/team-sync/group-mappings", func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, "PATCH")
-		fmt.Fprint(w, `{"groups": []}`)
-	})
-
-	input := IDPGroupList{
-		Groups: []*IDPGroup{},
-	}
-
-	groups, _, err := client.Teams.CreateOrUpdateIDPGroupConnections(context.Background(), "o", "s", input)
-	if err != nil {
-		t.Errorf("Teams.CreateOrUpdateIDPGroupConnections returned error: %v", err)
-	}
-
-	want := &IDPGroupList{
-		Groups: []*IDPGroup{},
-	}
-	if !reflect.DeepEqual(groups, want) {
-		t.Errorf("Teams.CreateOrUpdateIDPGroupConnections returned %+v. want %+v", groups, want)
-	}
-}
